+++ NEWS.md
@@ -15,6 +15,7 @@ current STUB variables.
 - Add check if keywords are in correct table.
 - Bugfix: `last_updated()` modified wrong table.
 - Bugfix: Remove NA values when creating px file (#205).
+- Bugfix: `px(input)` can be a path to an `.rds` file.
 
 # pxmake 0.10.0